Introduce IRoutingServiceInternal service.
